requirements.txt
# Core dependencies
neo4j>=5.14.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
python-dotenv>=1.0.0

# Oracle database connector (optional - only if using real Oracle)
# oracledb>=2.0.0

# Data processing
pandas>=2.0.0

# OpenAI
openai>=1.12.0

# Jupyter (optional, for notebooks)
jupyter>=1.0.0
ipykernel>=6.20.0

# Additional dependencies for better compatibility
typing-extensions>=4.5.0
pydantic>=2.0.0
